requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
chromadb
pdf2image
faiss-cpu
langchain_google_genai